refactor(api): move quest scripts to dedicated context
