+++ api/src/quest/scripts/update-name-of-combined-courses.js
@@ -1,8 +1,8 @@
-import { usecases } from '../../src/quest/domain/usecases/index.js';
-import { commaSeparatedNumberParser } from '../../src/shared/application/scripts/parsers.js';
-import { Script } from '../../src/shared/application/scripts/script.js';
-import { ScriptRunner } from '../../src/shared/application/scripts/script-runner.js';
-import { DomainTransaction } from '../../src/shared/domain/DomainTransaction.js';
+import { commaSeparatedNumberParser } from '../../shared/application/scripts/parsers.js';
+import { Script } from '../../shared/application/scripts/script.js';
+import { ScriptRunner } from '../../shared/application/scripts/script-runner.js';
+import { DomainTransaction } from '../../shared/domain/DomainTransaction.js';
+import { usecases } from '../domain/usecases/index.js';
 
 export class UpdateCombinedCoursesNameScript extends Script {
   constructor() {
