+++ api/tests/quest/unit/scripts/update-name-of-combined-courses_test.js
@@ -1,9 +1,9 @@
 import sinon from 'sinon';
 
-import { UpdateCombinedCoursesNameScript } from '../../../scripts/prod/update-name-of-combined-courses.js';
-import { DomainTransaction } from '../../../src/shared/domain/DomainTransaction.js';
-import { expect } from '../../test-helper.js';
-import { catchErr } from '../../tooling/test-utils/error.js';
+import { UpdateCombinedCoursesNameScript } from '../../../../src/quest/scripts/update-name-of-combined-courses.js';
+import { DomainTransaction } from '../../../../src/shared/domain/DomainTransaction.js';
+import { expect } from '../../../test-helper.js';
+import { catchErr } from '../../../tooling/test-utils/error.js';
 
 describe('UpdateNameOfCombinedCoursesScript', function () {
   let usecasesStub, logger, domainTransactionStub;
